pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mldata-cli"
version = "0.4.0"
description = "Unified CLI for ML dataset acquisition and preparation"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Noé Flandre", email = "noe@example.com"}
]
keywords = ["machine-learning", "datasets", "cli", "ml", "data"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "polars>=0.20.0",
    "duckdb>=0.9.0",
    "httpx>=0.25.0",
    "keyring>=24.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "diskcache>=5.6.0",
    "datasets>=2.14.0",
    "openml>=0.12.0",
    "kaggle>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "vcrpy>=5.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "types-pyyaml>=6.0.0",
]

[project.scripts]
mldata = "mldata.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/mldata"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.coverage.run]
source = ["mldata"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
